pyproject.toml
[tool.poetry]
name = "farmer"
version = "2.0.0"
description = "Auto Machine Learning"
authors = ["Hiroki Matsuzaki"]

[tool.poetry.dependencies]
python = "3.6.9"
scikit-learn = "^0.22.2"
dataclasses = "^0.7"
opencv-python = "^4.2.0.34"
pretrainedmodels = "^0.7.4"
albumentations = "^0.4.3"
optuna = "^2.0.0"
pandas = "^1.0.5"
tensorflow-addons = "0.11.2"
tensorboard-plugin-profile = "^2.3.0"
efficientnet = {git = "https://github.com/qubvel/efficientnet"}
h5py = "2.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
